+++ CHANGELOG.md
@@ -57,6 +57,7 @@
   - `solver.>?` (-> `.isGt`)
   - `solver.<=?` (-> `.isLte`)
   - `solver.<?` (-> `.isLt`)
+- Drop support for the `defaultDomain` option. The default is now a full domain (0 ... 1000000) and otherwise you'll have to explicitly declare it.
 
 ## v2.3.4: